+++ test/power.jl
@@ -2,11 +2,11 @@ using Test
 using ManifoldsBase
 using ManifoldsBase:
     AbstractNumbers,
-    DefaultManifold,
     ℝ,
     ℂ,
     NestedReplacingPowerRepresentation,
-    VectorSpaceType
+    VectorSpaceType,
+    DefaultManifold
 using StaticArrays
 using LinearAlgebra
 using Random
@@ -409,6 +409,16 @@ end
         N = PowerManifold(M, NestedPowerRepresentation(), 2, 3; parameter = :type)
         p = rand(N)
         @test zero_vector(N, p) == 0 .* p
+    end
+
+    @testset "TangentSpace" begin
+        M = ManifoldsBase.DefaultManifold(3)
+        N = PowerManifold(M, NestedPowerRepresentation(), 2)
+        p = [[1.0, 2.0, 3.0], [4.0, 5.0, 6.0]]
+        TpN = TangentSpace(N, p)
+        Tp1M = TpN[1]
+        @test base_point(Tp1M) === p[N, 1]
+        @test base_manifold(Tp1M) === M
     end
 
     @testset "fill" begin