+++ NEWS.md
@@ -5,11 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.15.10] 15/05/2024
+## [0.15.10] 19/05/2024
 
 ### Added
 
 * Functions `fill(p, N)` and `fill!(P, p, N)` to fill values into a point on a power manifold `N`.
+* introduce a `base_point(TpM)` to access the base point of a tangent space
+* introduce `TpM[i]` to access tangent spaces of factors from an `AbstractPowerManifold` or a `ProductManifold`.
+
 
 ## [0.15.9] 02/05/2024